ai-tools/mlx-model-hub/backend/pyproject.toml
[project]
name = "mlx-hub"
version = "0.1.0"
description = "Local-first MLX Model Hub for Apple Silicon"
readme = "README.md"
authors = [
    { name = "Saagar", email = "saagar210@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "pydantic-settings>=2.5.0",
    "mlflow>=2.18.0",
    "mlx>=0.22.0",
    "mlx-lm>=0.21.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-httpx>=0.49b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "prometheus-client>=0.21.0",
    "sse-starlette>=2.2.0",
    "asyncpg>=0.30.0",
    "httpx>=0.27.0",
    "psutil>=6.0.0",
    "greenlet>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "aiosqlite>=0.20.0",
]

[project.scripts]
mlx-hub = "mlx_hub.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mlx_hub"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
